requirements.txt
django
python-decouple
dj-database-url
requests


gunicorn
# for database
psycopg2

# R2 storages
django-storages[s3]